页面消失时，取消网络请求

+++ HuaHong/数据存储/线上存储/LeanCloudViewController.m
@@ -110,7 +110,7 @@
         }
         
         
-        NSLog(@"count: %ld",objects.count);
+        NSLog(@"count: %lu",(unsigned long)objects.count);
     }];
 }
